Restructured and cleaned up the code base.
