+++ Cake.Frosting/Abstractions/IFrostingContext.cs
@@ -2,15 +2,16 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using Cake.Core.Composition;
+using Cake.Core;
 
+// ReSharper disable once CheckNamespace
 namespace Cake.Frosting
 {
     /// <summary>
-    /// Represents a collection of service registrations.
+    /// Represents the build context.
     /// </summary>
-    /// <seealso cref="ICakeContainerRegistrar" />
-    public interface ICakeServices : ICakeContainerRegistrar
+    /// <seealso cref="ICakeContext" />
+    public interface IFrostingContext : ICakeContext
     {
     }
 }
