+++ Cake.Frosting/CakeHost.cs
@@ -14,9 +14,6 @@ namespace Cake.Frosting
 {
     internal sealed class CakeHost : ICakeHost
     {
-        // ReSharper disable once NotAccessedField.Local
-        private readonly ILifetimeScope _scope;
-
         private readonly CakeHostOptions _options;
         private readonly IFrostingContext _context;
         private readonly IEnumerable<IFrostingTask> _tasks;
@@ -28,27 +25,33 @@ namespace Cake.Frosting
         private readonly CommandFactory _commandFactory;
         private readonly EngineInitializer _engineInitializer;
 
-        public CakeHost(CakeHostOptions options, CakeHostServices services, ILifetimeScope scope,
-            IFrostingContext context,
+        // ReSharper disable once NotAccessedField.Local
+        private readonly ILifetimeScope _scope;
+
+        public CakeHost(IFrostingContext context, ICakeEnvironment environment, ICakeEngine engine, ICakeLog log,
+            ILifetimeScope scope, CakeHostOptions options, EngineInitializer engineInitializer, CommandFactory commandFactory,
             IEnumerable<IFrostingTask> tasks = null, IFrostingLifetime lifetime = null, IFrostingTaskLifetime taskLifetime = null)
         {
-            Guard.ArgumentNotNull(scope, nameof(scope));
-            Guard.ArgumentNotNull(services, nameof(services));
             Guard.ArgumentNotNull(context, nameof(context));
+            Guard.ArgumentNotNull(environment, nameof(environment));
+            Guard.ArgumentNotNull(engine, nameof(engine));
+            Guard.ArgumentNotNull(log, nameof(log));
+            Guard.ArgumentNotNull(scope, nameof(scope));
             Guard.ArgumentNotNull(options, nameof(options));
+            Guard.ArgumentNotNull(engineInitializer, nameof(engineInitializer));
+            Guard.ArgumentNotNull(commandFactory, nameof(commandFactory));
 
             _options = options;
+            _engineInitializer = engineInitializer;
+            _commandFactory = commandFactory;
             _scope = scope; // Keep the scope alive.
             _context = context;
+            _environment = environment;
+            _engine = engine;
+            _log = log;
             _tasks = tasks;
             _lifetime = lifetime;
             _taskLifetime = taskLifetime;
-
-            _environment = services.Environment;
-            _engine = services.Engine;
-            _log = services.Log;
-            _commandFactory = services.CommandFactory;
-            _engineInitializer = services.EngineInitializer;
         }
 
         public int Run()
