+++ Cake.Frosting/Extensions/CakeHostBuilderExtensions.cs
@@ -5,6 +5,7 @@
 using Cake.Frosting.Internal;
 using Cake.Frosting.Internal.Arguments;
 
+// ReSharper disable once CheckNamespace
 namespace Cake.Frosting
 {
     /// <summary>
